refactor: extract ActionBar component
