+++ api-editor/gui/src/app/App.tsx
@@ -39,7 +39,7 @@ import {
     selectShowPackageDataImportDialog,
     toggleIsExpandedInTreeView,
 } from '../features/packageData/packageDataSlice';
-import SelectionView from '../features/packageData/selectionView/SelectionView';
+import {SelectionView} from '../features/packageData/selectionView/SelectionView';
 import TreeView from '../features/packageData/treeView/TreeView';
 import { useAppDispatch, useAppSelector } from './hooks';
 import PythonFunction from '../features/packageData/model/PythonFunction';
